Credit calculation login in work.

+++ BankClient/DAL/Entities/Credit.cs
@@ -8,7 +8,7 @@ namespace DAL.Entities
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int PercentRate { get; set; }
+        public double PercentRate { get; set; }
         public double MinSum { get; set; }
         public double MaxSum { get; set; }
         public int MinMonthPeriod { get; set; }
